Bail out of typescript
